# documentation/mmda-openapi.yaml
swagger: "2.0"
info:
  description: "MMDA Backend API"
  version: "1.0.0"
  title: "MMDA API"
  termsOfService: "http://www.linguistik.uni-erlangen.de/"
  contact:
    email: "stefan.evert@fau.de"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "linguistik.uni-erlangen.de"
basePath: "/api"
schemes:
  - "http"
  - "https"
...
paths:
...
  /user/{userID}/discourseme:
    get:
      description: "Get User Discoursemes"
      operationId: "getDiscoursemes"
      tags:
        - "discourseme"
      responses:
        200:
          description: "List of Discoursemes"
    post:
      description: "Generate a new Discourseme object"
      tags:
        - "discourseme"
      operationId: "postDiscourseme"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: "name"
        name: "name"
        description: "Name of Discourseme"
        required: true
        type: "string"
      - in: "items"
        name: "items"
        description: "Lexical Items"
        required: true
        type: "array"
      responses:
        201:
          description: "Discourseme ID"
  /user/{userID}/discourseme/{discoursemeID}:
    get:
      description: "Get Discoursemes Details"
      operationId: "getDiscoursemeDetails"
      tags:
        - "discourseme"
      responses:
        200:
          description: "Discourseme Details"
        404:
          description: "No such Discourseme"
    put:
      description: "Update Discoursemes Details"
      operationId: "putDiscoursemeDetails"
      tags:
        - "discourseme"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: "name"
        name: "name"
        description: "Name of Discourseme"
        required: true
        type: "string"
      - in: "items"
        name: "items"
        description: "Lexical Items"
        required: true
        type: "array"
      responses:
        200:
          description: "Discourseme ID"
        404:
          description: "No such Discourseme"
        409:
          description: "Cannot edit topic Discourseme"
    delete:
      description: "Delete Discoursemes"
      operationId: "DeleteDiscourseme"
      tags:
        - "discourseme"
      responses:
        200:
          description: "Discourseme Deleted"
        404:
          description: "No such Discourseme"
        409:
          description: "Cannot delete topic Discourseme"
...